+++ src/checkers/file_present.rs
@@ -53,7 +53,7 @@ impl Check for FilePresent {
 
 #[cfg(test)]
 mod tests {
-    use std::{collections::HashSet, fs::File};
+    use std::fs::File;
 
     use super::*;
 
@@ -69,7 +69,7 @@ mod tests {
             file_to_check,
             file_type_override: None,
             file_with_checks,
-            tags: HashSet::new(),
+            tags: Vec::new(),
         };
 
         let file_present_check =
@@ -92,7 +92,7 @@ mod tests {
             file_to_check,
             file_type_override: None,
             file_with_checks,
-            tags: HashSet::new(),
+            tags: Vec::new(),
         };
 
         let file_present_check =
@@ -111,7 +111,7 @@ mod tests {
             file_to_check,
             file_type_override: None,
             file_with_checks,
-            tags: HashSet::new(),
+            tags: Vec::new(),
         };
 
         let mut placeholder_table = toml_edit::Table::new();
@@ -137,7 +137,7 @@ mod tests {
             file_to_check,
             file_type_override: None,
             file_with_checks,
-            tags: HashSet::new(),
+            tags: Vec::new(),
         };
 
         let mut placeholder_table = toml_edit::Table::new();