+++ src/checkers/file_regex.rs
@@ -92,7 +92,6 @@ impl Check for FileRegexMatch {
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashSet;
     use std::fs::File;
     use std::io::Write;
 
@@ -110,7 +109,7 @@ mod tests {
             file_to_check,
             file_type_override: None,
             file_with_checks,
-            tags: HashSet::new(),
+            tags: Vec::new(),
         };
 
         let mut check_table = toml_edit::Table::new();
@@ -152,7 +151,7 @@ mod tests {
             file_to_check,
             file_type_override: None,
             file_with_checks,
-            tags: HashSet::new(),
+            tags: Vec::new(),
         };
 
         let mut check_table = toml_edit::Table::new();