+++ src/checkers/lines_absent.rs
@@ -71,7 +71,6 @@ impl Check for LinesAbsent {
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashSet;
     use std::fs::File;
     use std::io::Write;
 
@@ -89,7 +88,7 @@ mod tests {
             file_to_check,
             file_type_override: None,
             file_with_checks,
-            tags: HashSet::new(),
+            tags: Vec::new(),
         };
 
         let mut check_table = toml_edit::Table::new();
@@ -124,7 +123,7 @@ mod tests {
             file_to_check,
             file_type_override: None,
             file_with_checks,
-            tags: HashSet::new(),
+            tags: Vec::new(),
         };
 
         let mut check_table = toml_edit::Table::new();
@@ -165,7 +164,7 @@ mod tests {
             file_to_check,
             file_type_override: None,
             file_with_checks,
-            tags: HashSet::new(),
+            tags: Vec::new(),
         };
 
         let mut check_table = toml_edit::Table::new();